Ignore GCC warning of unsafe floating point comparisons.
Due to optimizations, there are few cases where comparing with a constant
is needed, e.g. in operator*=. G++ will give us a warning for this.
Since we know what we are doing, surpress the warning.

The only drawback for this workaround is if somebody includes
qtransform.h in his code and disable the float-equal warning since
the warning will be activated again.

Reviewed-by: Samuel Rødal