Reintroduce the unaligned-unaligned 32-bit code that I had removed out of ignorance.
If both pointers are out of 4-byte alignment, doing the first load
will align them so we can do 32-bit comparisons. Lars's code had this
before, but I misunderstood it and removed, thinking it was doing
misaligned accesses.

I experimented with moving the tail comparison above the 32-bit
comparison to save a register, but it made things worse.

Reviewed-By: Bradley T. Hughes