Made a small correction and optmization to the event dispatcher.
By keeping record of the iteration count of the event loop, we can
determine whether the active object being processed is being run for
the first time or not in the current iteration. This is slightly
faster, since it saves us from adding the AO to the deferred queue
the second time around if the iteration is different. It is also
slightly more correct, since otherwise running a timer twice would
take three calls to processEvents instead of two (first time: run AO,
second time: add to queue, third time: run AO from the queue).

RevBy:    Trust me
AutoTest: Passed