Add an autotest to check that the network test server works
This test verifies that the network test server is behaving as
expected. I think I caught all the services we need testing in the
server, but there's still some work to be done:

1) verify that the FTP files are there where they're supposed to be
2) verify that FTP writable areas are writable
3) verify that the HTTP server has the correct files too
4) verify that the HTTP server requests credentials for the protected area
5) attempt to do NTLM authentication to verify the password (probably
   can't be done with netChat)
6) add Windows SMB tests (//qt-test-server/etc.)
7) add SQL tests (connecting to the SQL server ports)

It would be good as well if we could not use QtNetwork. If you break
QtNetwork, this test breaks too, so we don't know where the fault
is. However, rewriting networking code will add another source of bugs
(same for the NTLM authentication).

Reviewed-By: João Abecasis