Fix a crash where QCocoaWindow get events after its widget is dead
The invariant that QCocoaWindow's lifetime is contained in a QWidget is
simply not true.

A top-level QWidget gets associated with a QCocoaWindow (which is
reference counted). However, it can be the case that we've destroyed our
QWidget, the link is removed, the window is hidden, but the window still
gets an event. In that case we would crash with an eventual null pointer
access. However, we don't really need to do anything in this case, so
just call super and return.

Task-number: 253402
Reviewed-by: Morten Sørvig