Revert focus apparance on itemviews for X11
We changed this primarily for the mac as active appearance on
widgets in itemviews should depend on the window activation state and
not on the focus widget. It was explicitly added back for windows only
but has been reported as a bug on X11 as well so we might as well keep
it mac-only for now.

Reviewed-by: mortens