Don't draw an arrow for toolbuttons that are text only and have a menu.
I made this change to make things look a bit better on the mac. However,
a toolbutton with text only looks strange and as a bonus the arrow
covers some of the text, which no one wants. It seems that similar
consructions in Cocoa don't show the arrow for text button, so we
shouldn't either.

Task-number: 253339
Reviewed-by: Jens Bache-Wiig