Fix building on other platforms.
The order of things in SRC_SUBDIRS is important here because src_phonon
should be built before src_webkit and src_plugins because those
projects might link against Phonon. This should be fixed properly when
we merge into 4.6 because Phonon might depend on OpenGL for rendering
to a texture hence the comment.