Add properties to QGraphicsItem to change the transformations component
With the new properties it is possible to easily animate.

Setting a transform using setTransform is incompatible with thoses
properties.  Accessing thoses propeties if you set previously a
transform will give you the default values.

Acknowledged-by: Code made with Andreas. Documentation written with Thierry.
This still need a more in depth code review and documentation review.
But it is urgent to commit now because the Animation API integration depends on it.