Enabled the use of promoted widgets in new form templates.
The form builder used for preview of the templates did not implement
custom widget handling (FormBuilder::createCustomWidgets()).
On encountering a promoted widget, the widget factory
would then create a Widget Database entry specifying QWidget
inheritance (emergency fallback) which could not be overidden later
on.
To fix this, moved the code for handling custom widgets from
QDesignerResource to QSimpleResource and create a new
formbuilder class that handles it especially for the New Form Dialog
preview.

Task-number:  254282