Fix toolbutton text incorrectly clipped on windows
In 4.5 the toolbutton icon rendering was changed somewhat and the
bottom line of text for tool buttons icons with TextUnderIcon set is
incorrectly clipped on Windows. The style reserves only 5 pixels but
tries to use 6 pixels for text and icon spacing, hence we adjust the
text rect one pixel up. This should be safe considering the fact that
we have a margin on both sides of the icon already and avoids actually
moving the icon positioning.

Task-number: 252554
Reviewed-by: trond