If EGL fails to provide a valid Visual ID, try XRender for ARGBs
This patch is inspired by the "Fix ARGB visuals" patch in the Maemo
branch. Thanks go to the author of that patch (who isn't signed up to
Gitorious and thus can't be named - you know who you are! Thanks!!).

This patch should also fix ARGB visuals (even if they are supplied by
EGL) as such visuals require a colormap.