BT: Prevent crash in Designer when using a scroll wheel to change a property.
There was some strangeness happening here with parents, but the main
problem was the fact that wheel was getting sent to the focusframe and
not to the widget below. However, the focusframe has the "transparent
for mouse events" flag set and wheel events probably should be
transparent as well.

Task-number: 253539
Reviewed-by: Richard Moe Gustavsen
(cherry picked from commit e77f2e595d9b9a6f078f37894733c52bbcfeb695)