Fixed qmake writing targets into the root of the current drive when DESTDIR=./ and using Windows, nmake and shadow builds.
qmake would canonicalize the DESTDIR of "./" to "".  Then it would check
if the original DESTDIR ended with "/", and if so, append it to the new
DESTDIR, resulting in a DESTDIR of "/" - the root of the current drive.
Don't do that.

This bug doesn't occur with in-source builds because qmake detects that
the source and build directories are the same directory and replaces
the DESTDIR of "./" with "" before it reaches the buggy code.

Autotest: included
Reviewed-by: Lincoln Ramsay