Fix a crash in Phonon::EffectWidget
When creating a UI based on double types we only assigned the
control pointer in certain cases. This would crash because the
tooltip did not check for the pointer, but the real issue was that
we didnt assign the control in the first place.

Task-number: 249710
Reviewed-by: Richard Moe Gustavsen
(cherry picked from commit 3037e466ebf21aa4a47a722a1e2ff497690cbef5)