A fix for a bug 252798
AVKON will show blanker - this will allow applications to change their
layout and ensures that 'intermediate' states are not shown to the end
user. This blanker is a full screen window and this causes Qt
application to receive visibility change event (not visible) and then
Qt application deletes backing store. Finally, when blanker is
removed, Qt application receives another visibility change event
(fully or partially visible) and backing store is recreated, but the
original version does not ensure that window contents are valid. In
fact, because of orientation change event, window server generates
redraw request and it is handled before receiving that second
visibility change event. In QSymbianControl::Draw doesn't draw
anything.

By adding this update() call, we ensure that window will be drawn
correctly eventually.

Reviewed-by: Jason Barron <jason.barron@nokia.com>