Ensure a hierarchy of menus fade out together.
On Mac OS X, when you have a large hierarchies of menus and you select
the item at the end of the hierarchy. It will flash and then the rest
will fade out at the same time. Qt would do a phased approach which was
what no one expected. Introduce a QMacWindowFader class that can hold an
arbitrary number of qwidgets and then on command fade them all down
pased on the set duration. The API is a bit clumsy but is prefect for
this internal API.

Task-#: 251700
Reviewed-by: Richard Moe Gustavsen