Optimized QRasterPixmapData::metric.
Avoid needless DPI calculations when not needed, and simplify the DPI
calculations to the point where most of the operations disappear. Also
avoid calling QImage::metric() so that we don't need to go through two
switch statements and an extra function call just to get the width of a
pixmap.

Reviewed-by: Trond