Modify QDesktopWidget to emit proper signals on resize.
The resize behavior of QDesktopWidget was somewhat undefined on Symbian
because we weren't actually changing the size of the widget. This patch
fixes that and also implements the resizeEvent() function so that it
properly emits signals. We didn't call resize() to change the size here
because our implementation of setGeometry_sys() ignores the desktop
widget since it doesn't have a backing store and isn't a real window.

Task-number: 253930