Fixed a possible assert in QGtkStyle
We now set and unset the GTK_HAS_FOCUS flag on the same
painting call only if focus is set instead of resetting it
on each painting call. This is a tiny optimization but also kills
a possible assert on certain versions of Gtk+ (as reported with
Red Hat Enterprise Linux 5).

Task-number: 254614
Reviewed-by: denis