Fixed: Setting a border using stylesheet for QComboBox adds an unwated frame.
This was already fix. But there was still a frame if there was a
stylesheet on the applicaiton.

The reason is that the frame's constructor call the style for some
hints. And later the combobox will query the style again for the frame
hint, before the view get polished.
The problem is that the StyleSheetStyle will fill the css cache with
wrong information on the first call.

This is not visible if there is no stylesheet as in the constructor, the
widget style is still the default application stylesheet if there is no
global applicaiton stylesheet.

Task-number: 254589
Reviewed-by: jbache
BT:
(cherry picked from commit d0bc0a26f8ac4c2f02819c262b8aa7c3dd1cad3b)