BT: Support saving forms with resources more than once in Qt Jambi
When a form is loaded, Designer will set the path to the resource as
an absolute path rather than the path actually stored in the file. There
is code to work around this to make file paths relative when saving the
file later on, but no work around for Qt Jambi. So when saving Qt Jambi
forms, you would get an absolute path to the resource which contains the
location of the resource file (.jar or on disk) This of course breaks
the concept of resources, as the .jui file was no longer portable. The
fix is to special case Qt Jambi resources and set the relative path
when loading them.

The patch has no effect on regular Designer.

Task-number: 254621
Reviewed-by: Kai Koehne