Fix wrong sorting when using the QFileSystemModel with QTreeView
An optimization was made to the sorting of QFileDialog to sort only
the current root (meaning what the user see). This avoided slowness
when the model was big with lots of leafs. The problem here is
for the treeview, the root is always the same, we expands only
nodes. In that case, a recursive sorting is needed to ensure that
all expanded nodes are correctly sorted (and filtered). This will
be slower that's why i use an hidden flag in the d pointer to
deactivate the recursive sort for the QFileDialog.

Task-number:254701
Reviewed-by:olivier
BT:yes
(cherry picked from commit 8e4300e2866fd28881853509df6ff054e13f841b)