greatly speed up QTransform::mapRect() for projective transforms
The code so far was converting the rect to a painterpath, mapping
that one and then taking the bounding rect. It is actually sufficient
to simply map the four corners of the rectangle and take the bounding
rect of these four points even in the projective case.

Fix the MAP macro to be more correct for degenerated projective
transformations.

Reviewed-By: Samuel