force activation of minimized windows on Windows mobile
When pressing the <- key on a Windows mobile device, the window gets
a minimized event (no other soft keys behave like that). Restoring the
window via the app menu isn't possible, because the window get a
WM_ACTIVATE but its internal state is still minimized.
It makes sense to unminimize activated apps on Windows mobile.

Task-number: 254673
Reviewed-by: thartman