More re-factoring of Graphics View's update mechanism.
This time with a recursive approach of processing dirty items. I've kept
the previous approach using a dirty list, but the recursive one is now
the default. Use QT_GV_USE_DIRTY_LIST=1 to swap.

I've also cached the item's device transform in both cases so that we
can re-use it later when drawing the item.