QGraphicsAnchorLayout: Temporary destructor to avoid memory leaks
Currently the itemCenterConstraints that are created when an item
is added to the layout are never being deleted.

Ideally this should be done by the time the item is removed from
the layout, this requires some changes in the data structures used
though. As a temporary solution we keep all of them until the
layout is destroyed.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>