Force qRound64() to take a double on platforms where qreal is a float.
In Symbian qreal is defined as float. And in qround64 the parameter
was qreal. This qround64 is used (only) in qvariant::convert, when
converting from double to e.g. int. This caused overflow bug for some
(close to max int) values, because in convert chain double value
was casted to float.

Task-number: 250267
Reviewed-By: Samuel Rødal
Reviewed-By: Jason Barron