Disable keypad navigation for the autotests and document it.
On Symbian keypad navigation is enabled by default since there is no
reliable way to detect the presence of touch or not. Enabling this
feature though causes problems for many of Qt's autotests since they
were not written with this in mind. Many tests send QTest::key*()
events to a widget and subsequently test the state of the widget, but
with keypad navigation enabled, the behavior changes because the widget
does not have edit focus by default.

Reviewed-by: axis <qt-info@nokia.com>