Support for SSL symbol runtime resolving in Symbian.
The commit enables -openssl config in addition to -openssl-linked
in Symbian OS platforms. In Symbian OS the symbol names are stripped
away from DLLand symbols are are referenced via ordinal numbers.
This change resolves the correct ordinals for SSL symbols in Symbian OS.

The change is a bit hackish, but is currently only way to resolve
symbols at runtime in Symbian OS - if STDDLL targettype is not counted.
The problem of STDDLL target type is that it is not supported by S60 3.1
and in addition the SSL libraries in S60 are not currently build with
STDDLL target but DLL target.

This mechanism should work as long as BC for SSL symbols is kept.
And S60 has given binary promise, so we should not have problems...