Added the -showinternal flag to qdoc.
When you set -showinternal on the command line, qdoc will include
everything marked \internal in the documentation. This flag is not
very useful at the moment for two reasons: (1) It generates hundreds
of qdoc errors because most of the things marked with \internal don't
have any documentation anyway, or the documentation has other errors
in it that weren't being detected because of the \internal. (2) There
is a bus error toward the end, which I haven't tracked down yet. For
now, use -showinternal at your own risk.