BT: Fixed crash on Mac caused by erroneous handling of native focus events.
On Mac, a widget with a NoFocus policy could still get focus
(if only temporarily) as the result of a native focus event.

In particular, a line edit with a completer should
not lose focus (if only for a brief moment) as a result of the
completer popup being shown. This will for example cause an
item delegate to think that the user has navigated away from
the cell and delete the line edit as a result. This will in turn
cause the completer to access a null pointer.

Reviewed-by: Richard Moe Gustavsen
Task-number: 254456 and 254460