Change QDesktopServices::TempLocation on Mac to be the same as QDir.
In the past, TempLocation would return something like:
/private/var/folders/4k/4k97GBy2Ha46D3DAWSLbQE+++TI/TemporaryItems

Now it returns something like:
/var/folders/4k/4k97GBy2Ha46D3DAWSLbQE+++TI/-Tmp-

This isn't that different and it's a temporary location so it shouldn't
effect many people (and it will be gone once the system reboots anyway),
but the issue could be that TemporaryItems isn't there and that it's a
bit more "Carbon" to use TemporaryItems.  Also our own documentation
claims they are equivalent and I'm happy to do that.

I also enforced the Qt-style of if-statements and got rid of the extra
qualifiers to make the code look nicer.

Task-number: 253806
Reviewed-by: Morten Sørvig