Improve detection of monotonic timer support
Some older systems support monotonic timers, but not POSIX.1-2001. The
sysconf() function is documented to return -1 if a feature is not
supported, otherwise it returns a value that's dependent on it's
argument. For POSIX.1-2001 compliant system, the value is 200112L, but
on other systems it may just return 1, for example.

Because of this, we check for the -1 return value (feature missing),
instead of relying on a specific (positive) return value.