Fix sorting bug and ensure render functions work. Make direct default.
Mark the children list for sorting when the Z value for items changes.
Change the signature of the recursive draw function slightly so that
the expose region is optional, and ensure we don't intersect with this
region if it's not available.

This change also flips the direct painting so that the default is to
use the recursive approach.