Ensure small and mini headers aren't bigger than "large" ones.
We were only "constraining" headers if they were large, but ignoring the
small and mini ones. This resulted in the small and mini headers looking
bigger than their normal ones which looked strange. Since the headers
have no real small or mini size for a header, we'll make sure to
constrain them all the same.

Task-number: 221115
Reviewed-by: Richard Moe Gustavsen