A faster item discovery function for rectangles (recursive).
This function works much faster than the last one, but still it can be
much faster. The main expense right now it seems is the transform
calculations, and the item->collidesWithPath call. There's still much
to gain here.

This function does not make use of the BSP tree's fast lookup so it makes
lookups slower in (e.g.) the chip demo.