BT: Remove duplicate code and fix font parsing.
QFont has a feature that you can pass a comma-separated list and it will
walk through the list and match the font that it hits first. There's a
nice static function that X11 and Windows uses, but the Mac was using an
older copied version of it. This old version didn't handle quoting which
is what happens in the style sheet. So, using the same code makes
everything work well. As a bonus, Creator looks correct again.

Reviewed-by: Simon Hausmann
(cherry picked from commit ad46e77420449ede2cb6c1ea2a810a2614520db9)