BT:  qt demo takes 100% of the cpu on X11
The tick timer is always active, even if the moving gree Qt logo is not visible.
But the code that is supposed to pause it when the app loose the focus doesn't works if
the moving Qt logo is not visible.

Also the call to syncX makes Xorg takes lot of cpu.

It doesn't fix the fact that the timer is still running while the green logo is not visible,
but at least doesn't take the cpu anymore if qtdemo loose the focus.

Task-number: 255020
Reviewed-by: Richard Moe Gustavsen
(cherry picked from commit 259b65c2f5d736dd7f6d81b6390f54464dd5f183)