qdoc: Fixed constructor position in summary lists.
The constructors had been displayed in the left column of the
table in the summary list, because there was no type. It was
caused by a strange test of the match index for 0, which was
true for constructors and destructors, anything without a type.
I removed that test, since I couldn't figure out what it was
for. We might see problems elsewhere in the docs because of
this, so beware. I didn't see any, but that test must have
been there for some reason.