Fixes a crash in QDoubleSpinBox
Removing dubious intermediate detection code that also had a buffer
overflow.  The results were inconsistent and not dependable on.
Processing was inefficient and end value to user experience dubious.

Test cases that abused the former behaviour were changed to consider
input in an Intermediate where it was previously considered Invalid.

With this change, user input will mostly be considered in an
intermediate state, until it is effectively validated.

Task-number: 255019
Reviewed-by: Anders Bakken