Allow qmake to compile using it's .pro file.
This fixes a problem where qmake would not compile when it was built
from it's .pro file because this method uses the pre-compiled header.
This header was causing a compile error in qlocale.cpp because
qtextstream.h was included and this includes qlocale.h. The problem in
qlocale.cpp was that it uses a define called QLOCALE_CPP to enable
extra functions in the class declaration, but the pre-compiled header
was preventing the qlocale.h from being re-processed and therefore the
function was never compiled in.

Reviewed-by: Marius Storm-Olsen