fix silly typo
Yeesh.
The function worked for the common case of the argument being a plain
script object (obviously, otherwise this would have been discovered
sooner), but it would never pick the less expensive path when replacing
the QObject pointer of an existing proxy. And if you passed in a
QVariant proxy (now who would ever do something like that...?),
it would assert.