Document that textVisible is optional when it comes to styles.
No progress bars on the mac show text and it would be bad if we allowed
it. There's nothing stopping people from connecting the valueChanged()
signal to a slot and have a real label layed out correctly that actually
updates with the amount of time it takes to complete, etc. This is more
what they do on Mac OS X if they decide to show a label.