Compatibility fix for QGraphicsScene::sceneRectChanged()/sceneRect().
We have to keep the growingItemsBoundingRect up-to-date if there's no
scene rect. The only difference now is that sceneRectChanged will not
be emitted before entering the event-loop, but the documentation
only states it'll be emitted when the scene rect changes, so we
consider it harmless.

Makes tst_QGraphicsView::sceneRect_growing and
tst_QGrahicsScene::sceneRect happy.

Reviewed-by: Andreas