Introduce Symbian CFbsBitmap <-> QPixmap conversion functions.
Introduce functions for converting to and from a CFbsBitmap on Symbian.
Currently these functions unfortunately have to copy the data to the
internal QImage stored in the raster pixmap data backend :( Hopefully
we can improve this with a native Symbian pixmap backend in the future.

The autotest generates both small and large CFbsBitmaps of different
formats and then verifies that the color makes it into the QPixmap.
Currently the indexed image formats don't seem to work, but this is
most likely because we don't set a palette for them. The semi-trans
bitmaps seem to work fine (verified visually), but the current QCOMPARE
doesn't consider the fact that the color was alpha blended.

Reviewed-by: Sami Merilä