Removes odd artifact in the chip demo.
Calling repaint() instead of update() is bad when having multiple
views (which GV perfectly supports). The result is that e.g. when
moving a chip in the chip demo, there's a visible lag between
each view. It can also be a performance killer on QWS, where
the surface is locked for each repaint().

Instead of calling repaint() we call update() as before, but we also
make sure the updates are processed immediately.