Made configure work with older versions of /bin/sh.
Some older versions of the Bourne shell (/bin/sh) don't
support the ! operator for negating the exit code of a pipeline.
In particular, this is true on certain (all?) Solaris 9 platforms.
The patch replaces all occurrences of the ! operator with equivalent
(but slightly less elegant) expressions that test the $? variable instead.

Reviewed-by: msorvig