Add a callback so Qt updates its color stuff when the profile changes.
We've had a problem with a stale cache for color profiles this should
make things work well. We get the callback for each display whether it
needs it or not, but honesly I would rather that we update this a few
times too many when people change their display profile than not at all.

FWIW, this code is inspired from Apple's Tech Note TN2035.