Fixed text drawing on Windows in 16 bit mode.
There were several problems with antialiased text in 16 bit mode
under Windows. No gamma correction was done, yet we prepared the
cached glyphs for gamma correction. The mask format we rendered
the glyphs into was also set to the desktop depth, which implied
that information was lost and the text looked rather odd.

Reviewed-by: Samuel
BT: yes