Revert two of my commits, restoring the original fix for focus handling.
The fix that was introduced adds regressions, and I don't see a way we
can fix it without breaking someone elses code. So restoring the
original fix that just avoid a crash (autotest by Thierry is included).

Revert "Revert Avoid a crash when setting a focus in a widget
hierarchy containing"

Revert "Setting a focus on a widget hierarchy which contains both
visible and invisible widgets could cause a crash."

This reverts commit be833a4f25a8ec8c3dd7a8ac4fa4b0507c93e7ee.
This partially reverts commit 1a7da7096bbda17197738061902f4489af234bc0

Reviewed-by: Thierry Bastian
Reviewed-by: Prasanth Ullattil