's60theme' is a commandline tool which converts Carbide.ui themes into an intermediate binary format that can be read by the simulated QS60Style. So, for example Designer (standalone or in Carbide) will be able to display a realistic S60 Ui.
The intermediate binary format hashes of QPictures and QColors,
streamed to a QByteArray and compressed. The QS60Style could not load
the Carbide.ui theme directly because SVG handling is unfortunately
not part of QtGui, and would require a dependency on the QtSvg module.

Also, 's60theme' uses QWebkit to parse the SVG graphics (inspired by
Ariya's 'WebKit-based SVG rasterizer' labs post). QtSvg had some issues
with the SVGs that come with Carbide.ui.