Add ItemSendsGeometryChanges, replacing itemChangeEnabled().
This flag toggles whether we should send notifications for setPos,
setMatrix, and setTransform. It's off by default. Docs have been updated.
All autotests pass. This change also cleans up a bit so that we both
have readable code, and keeping the optimized path for when we need to
send the notifications.

By enabling this flag by default we are going to trigger regressions in
end-user code.

Reviewed-by: bnilsen